Some (big) modifications to paths.
First, validation is painfully slow, so I've added an option to validate less often and to validate on smaller sets.

Second, the initial h0 for the conditonal was all wrong and too complex. It is now the output of a network starting at h (the latent variable of the sbn)